requirements.txt
google-api-python-client>=2.166.0
google-auth>=2.38.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=0.4.1
openai>=1.68.2
python-dotenv>=1.1.0
six>=1.16.0